pyproject.toml
[project]
name = "sk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "python-dotenv>=1.2.1",
    "agent-framework",
]
